Moved REST API clients to api.client package, reverted to debian-based Python image, updating last_seen column in the physical_devices table.
The REST API clients will be required by each listening process so they
had to move from the ttn package.

Development will now be done on x86_64 linux boxes so the Python image
can go back to being debian-based to avoid the rust dependency that
slows down the alpine builds.

The ttn webhook now updates the last_seen column in the physical_devices
table, and takes the app_id and dev_id as webhook query parameters which
saves a few lines of code. Added a static decoder method to the
Location model so it can unpack the device location from a TTN device
definition.