Added a naive physical -> logical device mapping service.
This mapper reads the physical_timeseries message queue and looks for a
logical device already mapped to the physical device the message came
from.

If a logical device is found then the message is forwarded on to the
logical_timeseries queue with the logical device id added in.

There is nothing reading the logical_timeseries message queue yet.