Removed EXPOSE directive from docker image.
The docker documentation says the EXPOSE directive is informational and
does not cause the port to actually be exposed but docker-compose ps
showed the port in the ports section. It was not mapped to a host port
but I thought this looked misleading, like the port was opened for
listening in the container so I have removed it from the Dockerfile.