Fixed another SQL syntax error.
Still need to fix the update_logical_device function as has been done for
the update_physical_device.

Fixed a few bugs related to the addition of source_ids to the phyiscal
device.

Changed the logical mapper to stop TTN messages except from a test app
so the code is not reliant on uids. It felt wrong to be reading physical
device information into the UbidotsWriter code unless it was necessary
for determining a device label.