A first attempt at improving the mq code.
The mq client now has 3 classes - a connection, a tx channel, and
a rx channel.

A process can have as many channels as required, and each channel
can have a different exchange and queue (for rx channels).

The intent is to allow most of the processes to use fanout
exchanges to handle multiple listeners, while still allowing
some processes to have simple direct exchanges for internal
use.

All exchanges and queues are declared as durable.