The broker now supplies the physical device information to decoders as input.device.
The GreenBrain poller now exits cleanly when the users env var is empty.

Added a notes file for postgres to record useful information.

Removed the execute bit from the compose yml file attributes.

The TTN processor checks the locally decoded results against the decoded
results from the uplink converter if they both exist. This is a
temporary thing to check the local decoders are running correctly.