More logging changes, exposed port changes.
Set propagate to false for the cid logger, and moved the cid to after
the logging level so it lines up in docker-compose logs output.

The webhook and restapi ports are only exposed to localhost on the
assumption a reverse proxy is running locally to terminate TLS
connections. This will need a re-think if the reverse proxy runs on a
different host but that will also need other changes.