Fixed and improved the webapp, added an option to make a new user an admin in the CLI.
Fixed: Physical and logical device name/location changes were not working after authentication
was added.

Changed: The webapp map now shows logical devices. Added extra info to the device popup on the
map in the webapp, plus a link to Ubidots.

Changed: Made some of the web app code adhere more closely to the python way of things.

Added: A new flag to the CLI user add command to set the read-only flag to false.