Refactored to use the IoTa Pydantic object models.
Fixed a copy/paste bug in the logical device table.
Removed the project-specific types and brought the Pydantic models in. This caused the majority of the changes.
Added the note id into the device notes table so the notes can be sorted in chronological order rather than using the ts field which is updated when a note is edited.