Committing various SCMN-specific changes.
Production docker-compose file now specifies a host directory for the database data due to size constraints on the original filesystem.

A 0.3s sleep added to the Ubidots write data call to avoid breaking the API rate limit (also added in master).

A fix to command line argument checking in the CLI.

A fix to device last seen formatting and new initial map limits on the web app.

Upgraded leaflet library version in the web app.